requirements.txt
PySide6>=6.4.0
requests>=2.28.0
beautifulsoup4>=4.11.0
packaging>=21.0
openpyxl>=3.0.0
opencv-python>=4.5.0
mss>=9.0.0
onnxruntime>=1.14.0
numpy>=1.21.0
Pillow>=9.0.0
shapely>=2.0.0
pyclipper>=1.3.0